Add backbone library.
So, the idea is that all.js is the globbed together parts of all of the
other javascript files in the project. Unless explicitly stated, it will
import them in an unpredictable way. So, if libraries depend on
libraries, those should be explicitly imported before those dependent
libraries.

Since many libraries depend on jQuery and Underscore, we make sure to
explicitly include those before the others get a chance to load.